BMI270 - initial port Dusking;  WIP; Failing, first cpnenct static gyro, then softbrick no connect
