+++ src/main/drivers/accgyro/accgyro.h
@@ -43,7 +43,7 @@
 #define GYRO_HARDWARE_LPF_NORMAL       0
 #define GYRO_HARDWARE_LPF_1KHZ_SAMPLE  2
 
-#if defined(USE_GYRO_SPI_ICM42688P)
+#if defined(USE_GYRO_SPI_ICM42688P) || defined(USE_ACCGYRO_BMI270)
 #define GYRO_HARDWARE_LPF_EXPERIMENTAL 3
 #else
 #define GYRO_HARDWARE_LPF_EXPERIMENTAL 1
