+++ src/main/drivers/accgyro/accgyro_spi_bmi270.h
@@ -21,8 +21,69 @@
 #pragma once
 
 #include "drivers/bus.h"
+#include "drivers/exti.h"
 
+// BMI270 registers (not the complete list)
+typedef enum {
+    BMI270_REG_CHIP_ID = 0x00,
+    BMI270_REG_ERR_REG = 0x02,
+    BMI270_REG_STATUS = 0x03,
+    BMI270_REG_ACC_DATA_X_LSB = 0x0C,
+    BMI270_REG_GYR_DATA_X_LSB = 0x12,
+    BMI270_REG_SENSORTIME_0 = 0x18,
+    BMI270_REG_SENSORTIME_1 = 0x19,
+    BMI270_REG_SENSORTIME_2 = 0x1A,
+    BMI270_REG_EVENT = 0x1B,
+    BMI270_REG_INT_STATUS_0 = 0x1C,
+    BMI270_REG_INT_STATUS_1 = 0x1D,
+    BMI270_REG_INTERNAL_STATUS = 0x21,
+    BMI270_REG_TEMPERATURE_LSB = 0x22,
+    BMI270_REG_TEMPERATURE_MSB = 0x23,
+    BMI270_REG_FIFO_LENGTH_LSB = 0x24,
+    BMI270_REG_FIFO_LENGTH_MSB = 0x25,
+    BMI270_REG_FIFO_DATA = 0x26,
+    BMI270_REG_FEAT_PAGE = 0x2F,
+    BMI270_REG_FEATURES_0_GYR_GAIN_STATUS = 0x38,
+    BMI270_REG_FEATURES_0_GYR_CAS = 0x3C,
+    BMI270_REG_FEATURES_1_G_TRIG_1 = 0x32,
+    BMI270_REG_FEATURES_1_GEN_SET_1 = 0x34,
+    BMI270_REG_ACC_CONF = 0x40,
+    BMI270_REG_ACC_RANGE = 0x41,
+    BMI270_REG_GYRO_CONF = 0x42,
+    BMI270_REG_GYRO_RANGE = 0x43,
+    BMI270_REG_AUX_CONF = 0x44,
+    BMI270_REG_FIFO_DOWNS = 0x45,
+    BMI270_REG_FIFO_WTM_0 = 0x46,
+    BMI270_REG_FIFO_WTM_1 = 0x47,
+    BMI270_REG_FIFO_CONFIG_0 = 0x48,
+    BMI270_REG_FIFO_CONFIG_1 = 0x49,
+    BMI270_REG_SATURATION = 0x4A,
+    BMI270_REG_INT1_IO_CTRL = 0x53,
+    BMI270_REG_INT2_IO_CTRL = 0x54,
+    BMI270_REG_INT_LATCH = 0x55,
+    BMI270_REG_INT1_MAP_FEAT = 0x56,
+    BMI270_REG_INT2_MAP_FEAT = 0x57,
+    BMI270_REG_INT_MAP_DATA = 0x58,
+    BMI270_REG_INIT_CTRL = 0x59,
+    BMI270_REG_INIT_DATA = 0x5E,
+    BMI270_REG_GYR_CRT_CONF = 0x69,
+    BMI270_REG_IF_CONF = 0x6B,
+    BMI270_REG_ACC_SELF_TEST = 0x6D,
+    BMI270_REG_GYR_SELF_TEST_AXES = 0x6E,
+    BMI270_REG_OFFSET_6 = 0x77,
+    BMI270_REG_PWR_CONF = 0x7C,
+    BMI270_REG_PWR_CTRL = 0x7D,
+    BMI270_REG_CMD = 0x7E,
+} bmi270Register_e;
+
+extern int8_t bmi270CasFactor;
+
+// Contained in accgyro_spi_bmi270_init.c which is size-optimized
 uint8_t bmi270Detect(const busDevice_t *bus);
 bool bmi270SpiAccDetect(accDev_t *acc);
 bool bmi270SpiGyroDetect(gyroDev_t *gyro);
+
+// Contained in accgyro_spi_bmi270.c which is speed-optimized
 uint8_t bmi270InterruptStatus(gyroDev_t *gyro);
+bool bmi270AccRead(accDev_t *acc);
+bool bmi270GyroRead(gyroDev_t *gyro);
